Add endpoints related with webapp

+++ src/models/price.js
@@ -4,6 +4,7 @@ import { removeID, nextID } from './_helper.js'
 const priceSchema = new mongoose.Schema({
   id: { type: Number, default: 1, required: true },
   product_name: { type: String, default: '', required: true },
+  product_title: { type: String, default: '', required: true},
   price: { type: Number, required: true },
   ts: { type: Number},
 })
@@ -23,9 +24,10 @@ priceSchema.set('toObject', removeID())
 
 const ProductPrice = database.model('price_requests', priceSchema, 'price_requests')
 
-async function SetDefaultPrice(productName, price) {
+async function SetDefaultPrice(productName, title, price, isRemoveAll) {
+  if (isRemoveAll)
     await ProductPrice.deleteMany({product_name: productName})
-    await ProductPrice.create({product_name: productName, price: price})
+  await ProductPrice.create({product_name: productName, product_title: title, price: price})
 }
 
 export { SetDefaultPrice, ProductPrice }
